requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-filter==2.2.0
django-heroku==0.3.1
-e git+https://github.com/gbozee/django-paystack.git@fd2d44767e4c3531add523181eda8001c5ef6170#egg=django_paystack
django-storages==1.9.1
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Paperboy==1.0.1
Pillow==7.0.0
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-async==0.6.2
rfc3986==1.4.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
yapf==0.29.0
